+++ packages/EasyApiPlatform/src/Common/IriConverter/IriConverterTrait.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyApiPlatform\Common\Filter;
+namespace EonX\EasyApiPlatform\Common\IriConverter;
 
 use ApiPlatform\Api\IriConverterInterface as LegacyIriConverterInterface;
 use ApiPlatform\Metadata\IriConverterInterface;
@@ -10,13 +10,13 @@ use ApiPlatform\Metadata\IriConverterInterface;
  * @deprecated Remove when API Platform 3 support is dropped
  */
 
-if (\interface_exists(LegacyIriConverterInterface::class)) {
+if (\interface_exists(LegacyIriConverterInterface::class) === false) {
     /**
      * @mixin \EonX\EasyApiPlatform\Common\Filter\AdvancedSearchFilter
      */
     trait IriConverterTrait
     {
-        protected function getIriConverter(): LegacyIriConverterInterface|IriConverterInterface
+        protected function getIriConverter(): IriConverterInterface
         {
             return $this->iriConverter;
         }
@@ -27,7 +27,7 @@ if (\interface_exists(LegacyIriConverterInterface::class)) {
      */
     trait IriConverterTrait
     {
-        protected function getIriConverter(): IriConverterInterface
+        protected function getIriConverter(): LegacyIriConverterInterface|IriConverterInterface
         {
             return $this->iriConverter;
         }
